Mask.to_array: Return (None, bounding_box) if mask is all-white
The bounding box is sufficient in that case; bitwise anding with the
mask would be a waste of time.

Here I'm assuming that the mask is either 0 or 255. Anything non-zero
will be treated as 255 (but only for this optimisation). It looks like
we don't normalise the pixels anywhere else, which surprised me. I'll
address that in a separate commit.