stbt-debug: Deduplicate debug images on disk
Debugging in the object repository can be slow because the node has to
upload all the images before it can be displayed in the web-ui, and when
the client clicks between different debug html pages identical images must
be downloaded again and again.  Instead we store them by (xxhash) digest
so they're naturally deduplicated.