HdmiCecControl: Don't call `RescanActiveDevices()` on startup
With libcec4 I'm seeing this take much longer than with libcec3. I don't
think it's necessary at startup, but if a keypress fails then we still
force the full rescan.

In case this has undesirable side effects with some devices, I have made
this configurable. Configure the control with "hdmi-cec::::rescan" to
force the full rescan on startup (i.e. the old behaviour).