load_image: Normalize alpha channel when loading numpy arrays
Passing a numpy array now behaves the same as passing a PNG.

I have to make a copy of the array because we modify the underlying
data, and we don't want to surprise the user. The array could be
read-only, too. The copy is faster than checking whether we need to do
the copy. Measuring on my laptop with a 720p image:

    >>> a = numpy.ones((720, 1280, 4), dtype=numpy.uint8)
    >>> chan = a[:, :, 3]
    >>> %timeit a < 255
    549 µs ± 6.25 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
    >>> %timeit numpy.any((a < 255) & (a > 0))
    4.17 ms ± 51.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
    >>> %timeit a.copy()
    311 µs ± 22.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

The actual operation takes 1.6ms so it might be worth adding caching at
some point:

    >>> %timeit chan[chan < 255] = 0
    1.56 ms ± 5.64 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

But I won't add caching until this becomes a common use-case. Numpy
arrays can be mutable so it's tricky. (I suppose files on disk can be
mutable too, but we don't worry about them changing from call to
call...)