Make BGRDiff the default for press_and_wait / detect_motion / wait_for_motion
The previous default (GrayscaleDiff) doesn't detect changes that are
different in colour but of a similar overall intensity, such as the
XFinity keyboard shown in `tests/test_diff.py::test_bgrdiff`.

Note that the `noise_threshold` has changed in scale (it's now 0-255
instead of 0.0-1.0) and in direction (a smaller number means a stricter
threshold). The new default is 25, so it's slightly stricter (the old
threshold of 0.84 would correspond to roughly a value of 40 in the new
algorithm).

To continue using the previous algorithm, run the following code early
in your test script (for example at the top level of your package's
`__init__.py`):

    stbt.press_and_wait.differ = stbt.GrayscaleDiff
    stbt.detect_motion.differ = stbt.GrayscaleDiff