Mask.from_alpha_channel: Don't raise if image doesn't have alpha channel
Instead, create a mask that "masks in" the whole image (i.e. a
region-based mask the size of the whole image). This is the size of the
image instead of `Region.ALL` because we don't know how it'll be used,
so we don't want to lose that information (potentially the caller could
translate it within a larger image?).