match: Always convert frame to 3 channels
Previously if `frame` was greyscale any alpha channel in `image` was
ignored.  Now the alpha channel is used if present.

Technically this is a breaking change for greyscale frames with color
images.  But I think it's unlikely to affect any existing tests as
greyscale frames are only created as a result of explicit image processing.