Refactor: Replace subclassing `FrameDiffer` with subclassing `Differ`
Instead of subclassing `FrameDiffer` we now subclass `Differ` and pass an
instance of that subclass to `FrameDiffer`.  So before you'd write:

    press_and_wait.differ = BGRDiff

and now it's

    press_and_wait.differ = BGRDiff()

This means that we can customise differ parameters by passing arguments to
the constructor like:

    press_and_wait.differ = BGRDiff(threshold=60)

To support this differs are now immutable.  The `previous_frame` state that
used to live in `BGRDiff` and `GreyscaleDiff` is now implemented in
`FrameDiffer` which is now a concrete class.

We still need to support passing differ parameters directly to
`press_and_wait`, etc. like `press_and_wait(min_size=...)`, so there's a
`.replace` argument which returns a new differ, but with those parameters
set.

`Differ` is designed so we can efficiently implement `GrayscaleDiff`
without doing the colour -> grayscale conversion on the same image twice -
just as we could before.  To avoid internal state the interface makes
preprocessing explicit.  This makes the classes harder to use directly, but
that's ok because users won't be using them directly.

Currently we don't override `preprocess_mask` anywhere, but it seems to
fall out naturally from the design.