BGRDiff: Implement inner loop in C
This removes a bunch of large temporary arrays from the BGRDiff hot path.
From looking at godbolt it looks like it vectorises well too.

Now a full-frame comparison takes <1 frame on the 1st gen Stb-tester Node
hardware, whereas before it was taking ~4.  This is important as we've
made BGRDiff the default now.