Drop support for PDU types "pdu" and "ipp"
...deleting `_ShellOutlet` and `stbt-power.sh`.

These are the two PDU types still implemented in shell.  I don't trust that
implementations are safe WRT injecting shell commands in `$hostname`, and
they include hard-coded usernames and passwords.

Both implementations are implemented in terms of `curl` and then using
`awk`, `grep` and `sed` on the result.  It would be easy enough to
re-implement in Python using `requests` if only there were tests with
example inputs and outputs.  If there are any users hopefully they will
get in touch and then we or they can reimplement this in Python
**with tests**.

As it is this is too hard to maintain and too risky so I'm removing it now.