Keyboard.add_grid: Merge identical adjecent keys
For example: this keyboard has a 3 square wide space bar and a two square
wide OK button on the bottom row:

    KEYBOARD = stbt.Keyboard()
    KEYBOARD.add_grid(
        stbt.Grid(
            stbt.Region(377, 476, 527, 228),
            data=[
                "qwertyuiop",
                "asdfghjkl,",
                ["CAPS"] + list("zxcvbnm.") + ["BACKSPACE"],
                ["?123"] + list("◀▶   -_") + ["OK"] * 2,
            ]
        ),
        mode="lower", merge=True
    )

Without this change you'd have to separately add space and ok as separate
keys and explicitly list the transitions up, left and right.  This is much
less verbose and more readable.