imgproc_cache: Limit cache size to 90% of filesystem size
I'm attempting to debug an issue where we get SIGBUS during `_cache_put`.
SIGBUS can happen when attempting to fault in a page when there is
insufficient space on the backing store to allocate one.  This should help
prevent that situation, and we should get a `MapFullError` instead.

Instead of just warning we now just delete the cache if it's full to
resolve any issues like this.  It's safe to do so while the database is
open, and the database will just be re-created when the next test-run
starts.