Expose `Transition` (return type of `press_and_wait`) as public API
Needed for type annotations on navigation functions etc.

I have called it "Transition" instead of "TransitionResult" because
having "TransitionResult" and "TransitionStatus" is too confusing. Now
the only confusion is limited to the internal implementation details in
`transition.py` where we have "Transition" vs. "_Transition" (the latter
is a private helper class).