Keyboard: Suport Page Object with `focus` property instead of `selection`
"focus" is going to be our official name going forward; "selection" is
ambiguous because you can have "selected" something and then move the
"focus" somewhere else without changing the active "selection".

I'm adding this to Keyboard for consistency with `navigate_1d`.

For backwards compatibility we use the `selection` property, if it
exists (even if `focus` also exists, to avoid changing behaviour for
existing users, in case they have both a `focus` and a `selection`, even
though it seems unlikely... but doing it this way means we can release
this in a v34 minor release).