ocr_eq.replacements: Custom __setitem__
There's a subtlety when adding your own replacements, if the replaced
value is itself already a key in the replacements. Previously, you would
have to do this:

    ocr_eq.replacements["İ"] = "I"                    # incorrect!
    ocr_eq.replacements["İ"] = ocr_eq.normalise("I")  # correct

Now, `ocr_eq.replacements["İ"] = "I"` is correct.

If a user replaces `ocr_eq.replacements` wholesale with a new dict,
it'll lose this behaviour; but then that's their problem. Either replace
it once with a single dict that you don't modify, or add to our one.

Note: I subclass `collections.UserDict` instead of subclassing `dict`
directly, so that `copy()` preserves my subclass type.