keyboard: Fix shortest path to ambiguous target
This is an edge case that we found when navigating the Channel 4 app on
Apple TV. There are 3 keyboard modes, and you press KEY_PLAY to switch
between them. The exact key you land on can vary based on previous
state, so to simplify our model we just said that pressing KEY_PLAY can
land you on *any* key of the next mode. This resulted in buggy
behaviour: `Keyboard.navigate_to("BACKSPACE")` would press
`KEY_PLAY` (over and over, until `navigate_timeout`).

Here's what was happening: BACKSPACE is available in all modes, so
`navigate_to` was calling `_keys_to_press` with 3 possible targets. We
do take into account the weights when calculating the shorting path to
any one of those targets, but we weren't taking into account the weights
when choosing which of those 3 paths was the shortest. Because of our
lazy hack when modeling the keyboard, it thought that a single
`KEY_PLAY` might get us to BACKSPACE in the next mode.

The fix is: It should always prefer a deterministic path over a
non-deterministic path† no matter how short the latter is.

† That's what we call a path in our model where pressing a key might
land you on any one of several states.

Note: I have committed the screenshots so you can see what the keyboard
looks like, but I don't actually use the screenshots in the test.