stbt.Keyboard: Allow disjoint modes
If there are two "A"s in the keyboard, it should succeed if it can find
a path to one of them.

For convenience I'm modelling a Keyboard that has 2 different
appearances (on different STBs) as a single Keyboard with 2 different
modes (and a single Page Object that handles both appearances). Both
modes have all the letters, and there's no way to get from one mode to
the other. This way I don't have to make my Page Object remember which
`Keyboard` instance to use — there's only one `Keyboard` instance.