Keyboard: Performance optimisation for `_find_keys`
This fixes a quadratic performance when adding keys, because each time
we added a key, we scanned through the entire list of keys to check that
it wasn't already present.

Note that we can't just look up the key by hash in the `G.nodes()` dict,
because we need to check that the region doesn't overlap with any
existing keys.

This speeds up `test_keyboard_definition_performance` from 0.4s to 0.25s
on my laptop.