CI: Run pyright type checker
pyright is written in TypeScript. The PyPI package is a
community-maintained Python wrapper / installer. In the dockerfile I
have to run it once because it downloads the npm packages on first use
(even with --version).

I have added most files to the "ignore" config, so that we can run this
in CI and avoid regressions in the files that *are* type-checked. The
idea is to remove these one by one as and when.

I have configured it to "basic" mode; the default is "standard", which
enables some extra checks such as "reportPossiblyUnboundVariable". We
can enable these when we fix the issues. "standard" would report only 22
additional issues (on the files not currently ignored) so I would
recommend fixing those first and enabling "standard" mode, before we
start fixing other files.