CH9329: Use `Serial.reset_input_buffer()` before making any requests
We've seen errors like:

> Invalid header in frame 0057ab008108.  Expected 0x57ab, got 0x0057

It's clear that a rogue 0x00 has sneaked into the serial stream, as the
0x57ab header is still there, but shifted over by one byte.  Once we get
in this state we don't recover as we're always off-by-one.

With this implementation we clear the buffer before any request, so
if there is extra noise on the line it will be cleared away.  There's
still the possibility that the noise could come after the clearance, but
before the response, but that's quite a narrow window, and we'll recover
on the next press anyway.